feat: add output generation for computed fields in terraform modules
