+++ pyproject.toml
@@ -60,8 +60,8 @@ Source = "https://github.com/EspenAlbert/atlas-init"
 path = "atlas_init/__init__.py"
 
 [tool.pytest.ini_options]
-addopts = "-vv --tb=short --log-cli-level=INFO --ignore=.venv/* --capture=no"
-# addopts = "-vv --tb=short --log-cli-level=INFO --ignore=.venv/* --capture=no --force-regen"
+# addopts = "-vv --tb=short --log-cli-level=INFO --ignore=.venv/* --capture=no"
+addopts = "-vv --tb=short --log-cli-level=INFO --ignore=.venv/* --capture=no --force-regen"
 testpaths = ["test_atlas_init"]
 asyncio_default_fixture_loop_scope = "function"
 filterwarnings = [
